install docusaurus-plugin-sass and include in the docusaurus.config.js file, set customCss property  to point to the SCSS file, rename all css files including the custom.css file to .scss and update all component js files import statements to reference the new .scss files
