+++ package.json
@@ -54,5 +54,8 @@
     "shell-utils": "1.x.x",
     "unified": "^10.1.0"
   },
-  "version": "20.1.2"
+  "version": "20.1.2",
+  "dependencies": {
+    "docusaurus-plugin-sass": "^0.2.3"
+  }
 }
