+++ website/src/components/CustomBanner/StandWithUkraine.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './StandWithUkraine.module.css';
+import styles from './StandWithUkraine.module.scss';
 import Link from '@docusaurus/Link';
 
 const StandWithUkraine = (props) => {
