+++ website/src/components/Homepage/Features/HomepageFeatures.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './HomepageFeatures.module.css';
+import styles from './HomepageFeatures.module.scss';
 import FeatureList from './FeatureList';
 import Features from './Features';
