+++ website/src/components/Homepage/Header/HomepageHeader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import styles from './HomepageHeader.module.css';
+import styles from './HomepageHeader.module.scss';
 
 const HomepageHeader = () => {
   const { siteConfig } = useDocusaurusContext();
