+++ website/src/components/Homepage/SubHeader/HomepageSubHeader.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './HomepageSubHeader.module.css';
+import styles from './HomepageSubHeader.module.scss';
 
 const HomepageSubHeader = () => {
   return (
