+++ website/src/components/Showcase/Hero.js
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import React from 'react';
 import Link from '@docusaurus/Link';
-import styles from './Hero.module.css';
+import styles from './Hero.module.scss';
 
 export default function ShowcaseHero() {
   return (
